Refactored additional files
