+++ packages/retail-ui-extensions/src/extension-api/cart-api/index.ts
@@ -1,8 +1,1 @@
-import type {CartApiContent} from './cart-api-content';
-
-/**
- * Access the merchant’s current cart.
- */
-export interface CartApi {
-  cart: CartApiContent;
-}
+export type {CartApiContent, CartApi} from './cart-api';
