+++ packages/retail-ui-extensions/src/extension-api/locale-api/locale-api.ts
@@ -5,3 +5,10 @@ export interface LocaleApiContent {
   /** Callback when user changes locale during runtime */
   setOnChange(onChange: (locale: string) => void): void;
 }
+
+/**
+ * Access the merchant’s current locale (in [IETF format](https://en.wikipedia.org/wiki/IETF_language_tag)) to internationalize your extension content.
+ */
+export interface LocaleApi {
+  locale: LocaleApiContent;
+}
